Updated READMEs and moved kinesis-analytics-v2 example into its own folder.
