+++ python/example_code/kinesis-analytics-v2/analytics_application.py
