+++ python/example_code/kinesis-analytics-v2/example.sql
