+++ python/example_code/kinesis-analytics-v2/requirements.txt
@@ -1,0 +1,2 @@
+boto3>=1.26.79
+pytest>=7.2.1
