+++ python/example_code/kinesis-analytics-v2/test/conftest.py
@@ -1,0 +1,13 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""
+Contains common test fixtures used to run unit tests.
+"""
+
+import random
+import sys
+
+# This is needed so Python can find test_tools on the path.
+sys.path.append("../..")
+from test_tools.fixtures.common import *
